Update patch set 2
Patch Set 2: Code-Review-1

Thank your very much for solving the "reload-button-to-the-right problem" (+2)

Sorry, but I do not like the new colors and the redesign of the reload-button.

The new color of the corpus statistic is in my opinion a little bit to loud(I mean "grell" in German and hope I found the right expression), that is kind of subjectiv.

But through the  redesign of the reload-button and the new colors of the inactive corpus statistic the inactivity and the reload button can now be easier overseen. I think it was user-friendlier before.

Patch-set: 2
Label: Code-Review=-1