Update patch set 2
Patch Set 2:

> Thank your very much for solving the "reload-button-to-the-right
 > problem" (+2)
 > 
 > Sorry, but I do not like the new colors and the redesign of the
 > reload-button.

I aligned the reload button with the first statistic line. It was unaligned completely. I disagree, that the reload button requires as much attention by the user as it had before. Though this can be discussed. The button would however definitely require to be aligned.

 > The new color of the corpus statistic is in my opinion a little bit
 > to loud(I mean "grell" in German and hope I found the right
 > expression), that is kind of subjectiv.

I agree - I ammended a slight change after comparing both color schemes. We should, probably, derive the lightest orange with a similar scheme, so they are balanced.

 > But through the  redesign of the reload-button and the new colors
 > of the inactive corpus statistic the inactivity and the reload
 > button can now be easier overseen. I think it was user-friendlier
 > before.

I don't think the outdatedness of the statistics justifies a complete new color-scheme, that does not fit in the overall scheme. The new color is more acceptable, so a user does not have the feeling that something is wrong that needs to be fixed. A mild indicator of outdatedness should be enough in my opinion.

Patch-set: 2
Label: Code-Review=0