Update patch set 3
Patch Set 3:

> I don't think the outdatedness of the statistics justifies a
 > complete new color-scheme, that does not fit in the overall scheme.
 > The new color is more acceptable, so a user does not have the
 > feeling that something is wrong that needs to be fixed. A mild
 > indicator of outdatedness should be enough in my opinion.
 > 
 > > I disagree, that the reload button requires as much attention by
 > the user > as it had before.
 > 
 > I still think it is better to make it a bit clearer for the user
 > than to mild.

Is your complaint more focussed on the more subtle reload button or the outdated statistics color? If you find a better fitting solution for the statistic with reusing existing colors, I am all for it. If you are fine with the statistics but want the reload button in the more engaging grey/orange-button scheme, this can be a compromise, though the button should not be larger than the other buttons and aligned with the statistics table.

Patch-set: 3
Reviewer: Gerrit User 19 <19@626241b7-b7d0-4197-b843-1ee23887649c>
Label: Code-Review=0