Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > (1 comment)
> > 
> > "beliebiger" suggests the misunderstanding that the user can decide.
> > 
> > My favorite is "willkürlicher" followed by "undefinierter"
> 
> Sorry, Marc, I oversaw your last comment. So I would change it now to "willkürlicher", unless Nils is highly objecting.

I understand that "beliebig" is also problematic, but "willkürlich" in my reading requires an authoritarian decision. Thats a) not the case and could b) mean, that the user believes the order is at least stable. (Btw. in linguistics "Willkürlichkeit" means something different as well.) So "undefiniert" or "unbestimmt" is in my opinion better.

Patch-set: 1