Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review+1
> > 
> > I thought the buttons "Search" and "Back" in the "Search"-Hint should also be switched?

I do not understand that. Right now, the "Back" Button comes first and behind it the Search Button, which defines the next step. Wouldn't it be confusing if the Back-Button comes after the Search-Button?

> And maybe the term "undefined" should also be used in the english translation instead of "arbitrary"."

I was not sure about that, because "arbitrary" includes the semantics of "beliebig" and "willkürlich" and that is probably what we want. I would match "zufällig" to "randomized"(we do not want that). 
But you are more corpus linguists than I am, so I bow my head to yours and Marc's experience and can change it accordingly.

Patch-set: 1
Reviewer: Gerrit User 33 <33@626241b7-b7d0-4197-b843-1ee23887649c>