Update patch set 1
Patch Set 1:

This seems to me to be 3 commits in one:

I would +2 the changes to server.js, result.js, and init.js, in case there was a spec. The changes seem valid and necessary for result panel plugins.

I would -2 the changes to the scss - I don't think plugins should require specific changes to the core css.

Regarding the demo: I don't know what it should do. An export widget demo should be part of the separate export service. When it's just there to demonstrate a widget in the result panel, it may be part of the plugin demo (with a more general name). Or is there something I am missing that is required and specific to this plugin?

Patch-set: 1
Reviewer: Gerrit User 19 <19@626241b7-b7d0-4197-b843-1ee23887649c>