Update patch set 8
Patch Set 8: Code-Review+2

I took a look at it. I think it's OK for the time being. The change of color in the result bar and the change of the search button makes it clear enough that the result is loaded.

Patch-set: 8
Reviewer: Gerrit User 33 <33@626241b7-b7d0-4197-b843-1ee23887649c>
Label: Code-Review=+2