Update patch set 1
Patch Set 1:

This is my first attempt for a publishable docker file for Kalamar. It's based on alpine and therefore pretty small (virtual image size ~342MB). It requires no perlbrew, so it's reasonable fast to build. Entrypoint is not yet defined and currently starts morbo only. Volumes for cache, secret and config are missing as well.

Build with:

  $ docker build . -t korap/kalamar

Run with:

  $ docker run -d -p 3000:3000 --name kalamar korap/kalamar

Enter image with:

  $ docker run -t -i korap/kalamar /bin/sh

Patch-set: 1
Reviewer: Gerrit User 19 <19@626241b7-b7d0-4197-b843-1ee23887649c>