Update patch set 1
Patch Set 1:

This is actually just a dirty fix for a problem that recently popped up: When a wide charater input is send to the server, the internal encoding can crash the system. This is independent from the handling of the LDAP server.

To mitigate this problem, some encodings are not allowed in this fix - but that's a bad approach tbh.

Patch-set: 1