Update patch set 4
Patch Set 4:

The required M::P::Piwik package is just in the release candidate phase on CPAN. It can already be installed from GitHub. I expect some results in the next days.

Currently CORS compliance is not the default, but it probably should be. But first this should be tested on instance/test.

Patch-set: 4