Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > A greek name for CI?!
> 
> I mean the fear of updating (or even touching) libraries 😊
>  

:D

> I think it very valuable to know if it runs with the latest library versions. If our deployment doesn't run, we can simply update all dependencies.

Sure - but the opposite may be true (sometimes as often): Your deployment runs but not on the latest versions. If this just means to update the code - fine, but if it means that a specific dependency doesn't work well with a specific dependency (like here some mojo plugins with the latest version of mojo), you cannot do much about but pin the versions for a while (or remove the dependencies). And this is not as easy in perl as it is in other languages. On the plus side: You won't forget to remove the pinning.

Patch-set: 1
Attention: {"person_ident":"Gerrit User 6 \u003c6@626241b7-b7d0-4197-b843-1ee23887649c\u003e","operation":"ADD","reason":"owner or uploader replied"}
Attention: {"person_ident":"Gerrit User 19 \u003c19@626241b7-b7d0-4197-b843-1ee23887649c\u003e","operation":"REMOVE","reason":"removed on reply"}